tacky: Refactoring massive match expression into smaller functions/modules.
